Update Rust in Dockerfile to 1.49.0
rclrs_examples fail to compile due to the following error
```
error[E0658]: use of unstable library feature 'renamed_spin_loop'
  --> /usr/local/cargo/registry/src/github.com-1ecc6299db9ec823/spin-0.9.2/src/relax.rs:26:9
   |
26 |         core::hint::spin_loop();
   |         ^^^^^^^^^^^^^^^^^^^^^
   |
   = note: see issue #55002 <https://github.com/rust-lang/rust/issues/55002> for more information

   Compiling bitvec v0.19.5
error: aborting due to previous error
```

As seen in the noted link, updating Rust can resolve the issue. This
updates Rust in the Dockerfile to 1.49.0, where the issue seems to be
resolved.