Flatten the module hierarchy
The modules for clients/services/publishers/subscriptions are moved up from the node/ directory
into the main directory.

This way, the node module only contains the implementation of the actual Node type.

Even though logically, a subscription for instance is a "part of" a node, the module layout should
not mirror this hierarchy, or we'd also move the node module below the context module and the
directories would become too nested.

This also mirrors the layout in rclcpp.