Incorporation of recommendations.
Unfortunately, I found the tutorial itself unfinished so far and wanted
to incorporate the justified criticism properly. One big change is that
there is now a separate `Build and Run` section, which generally
explains the correct building and starting of nodes. There are also a
few sections that deal with explanations and a few that most people skip
over. I'm largely satisfied so far and of course open to suggestions for
improvement.