Fix rclrs to compile after rebase
Update uses of a `Mutex<rcl_node_t>` with a `NodeHandle`, as was done
elsewhere in the crate. Also, correct the documented UUID size to 16
rather than 24.

The minimal_action_client compiles, but the minimal_action_server does
not yet, complaining about expecting the `rmw` versions of messages
rather than the idiomatic ones.