Handle action server/client readiness in WaitSet and executor
Currently, action servers and clients that are ready in multiple ways
are returned to the executor as a list of pairs, with one readiness mode
per entry. This could alternatively be encoded as a bitfield or similar
struct, with any given client/server only occurring once in the list.
However, to ensure that the executor has control over execution order,
we would need to expose individual `execute_readiness_mode()` methods
from the client and server, rather than a unified `execute(Mode)`
method. That's fine too, but something to keep in mind.