Add rcl_action error codes
There's one error code, ActionNameInvalid, that conflicts with the
EventInvalid code. We should consult with upstream about this, as it
causes issues for representing error codes as enum values. These two
error codes are probably never returned by the same function, but it
would be better to keep them unique.