[WIP] Start defining server/client execute functions
This is just the basic layout. I'm trying to avoid defining the `take_*`
functions that rclcpp has to link taking messages to executing on them.
I'm not sure if that's actually worthwhile yet.

This area should also be revisited once it's functional to see whether
portions can be moved into the non-polymorphic subfunctions. Doing so
could reduce compile times by avoiding excessive monomorphization.