Use rcl-allocated goal handle pointer in ServerGoalHandle
The `rcl_action_accept_new_goal()` function returns a pre-allocated
`rcl_action_goal_handle_t` pointer, which is also stored within the
action server proper. This means we cannot store a non-pointer version
of this in the `ServerGoalHandle`. Instead, we'll keep a mutex-guarded
mutable pointer. The `Arc` is unnecessary since this pointer is never
shared with anyone. Also, we need to clean up the goal handle by calling
`rcl_action_goal_handle_fini()` when the `ServerGoalHandle` is dropped.