Add UUID->GoalHandle hash-map to action server
This requires a mutex to enable simultaneous goal acceptance in a
multithreaded executor. We also make ServerGoalHandles implement Send
and Sync, since they will be accessed by multiple threads during
callbacks. Due to containing a raw pointer field, the type doesn't
automatically implement Send/Sync; however, we guarantee that the uses
of this pointer is safe and synchronized, allowing us to safely
implement the traits.